refactor(api): rename order to orders and delete dead create route
docs/conventions/route-naming.md asks collection paths to be plural and
keeps actions in the HTTP method rather than a segment of their own, so
/api/order and /api/order/create both read against the convention.

The create route turned out to have no callers at all: CheckoutForm
already posts through @/actions/createOrder, and nothing else in the
repository or the route constants references the endpoint. It is a
leftover from a migration that already completed, so delete it rather
than move it.

Rename the remaining collection route to /api/orders and update its one
consumer, the my-orders "load more" SWR key, along with the integration
test that mounts the handler.
