--- src/app/api/order/create/route.ts
@@ -1,53 +1,0 @@
-import type { APIRouteResponse} from "@/boundary";
-import { routeError } from "@/boundary";
-import { AppError } from "@/core/domain/error";
-import type { NextRequest } from "next/server";
-import { z } from "zod";
-import { validateAndFlatten } from "@/core/utils/validate-and-flatten";
-import { requireAuth } from "@/services/auth";
-const createOrderRequestSchema = z.object({
-  productId: z.string().min(1, { message: "상품 ID가 필요합니다." }),
-  originalPrice: z.number().positive({ message: "가격은 양수여야 합니다." }),
-  selectedOptions: z.array(
-    z.object({
-      _id: z.string(),
-      label: z.string(),
-      price: z.number(),
-    }),
-  ),
-  totalPrice: z.number().positive({ message: "총 가격은 양수여야 합니다." }),
-  buyerName: z.string().min(2, { message: "이름은 2자 이상이어야 합니다." }),
-  buyerEmail: z.string().email({ message: "유효한 이메일이어야 합니다." }),
-  buyerTel: z.string().regex(/^\d{3}-\d{3,4}-\d{4}$/, {
-    message: "연락처 형식이 올바르지 않습니다.",
-  }),
-});
-
-type CreateOrderRequest = z.infer<typeof createOrderRequestSchema>;
-
-export const POST = async (
-  req: NextRequest,
-): Promise<APIRouteResponse<{ merchantUid: string; orderId: string }>> => {
-  try {
-    await requireAuth();
-
-    const body = await req.json();
-
-    // 유효성 검사
-    const parsed = validateAndFlatten<CreateOrderRequest>(
-      createOrderRequestSchema,
-      body,
-    );
-
-    if (!parsed.success) {
-      throw new AppError("VALIDATION", "입력값을 확인해주세요", parsed.error);
-    }
-
-    // TODO: 이 API는 아직 완전히 구현되지 않았습니다.
-    // createOrderService를 사용하여 주문을 생성하고,
-    // payment는 결제 완료 후 syncPayment로 생성됩니다.
-    throw new AppError("DISABLED", "이 API는 현재 사용할 수 없습니다.");
-  } catch (error) {
-    return routeError(error);
-  }
-};
